+++ client/components/index.js
@@ -8,4 +8,4 @@ export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
 export {default as Single_Song} from './single-song'
-export {default as CartContainer} from './cart'
+export {default as Cart} from './cart/index'
